Fix OnPageNotFound
I'm not sure why you used REQUEST_URI here. At least in multi-context setups this doesn't work since it would contain the base_url of the context and so no matching resource would be found (since the uri of the resource does not contain the base_url).

So like in #22 I'm using $_REQUEST which works fine.